Final_swarm/backend/requirements.txt
Flask
flask-cors
psycopg2-binary
python-dotenv
